+++ pkg/datapath/pol.go
@@ -2,12 +2,27 @@ package datapath
 
 /*
 #include <string.h>
+#include <time.h>
+
+unsigned long long get_os_usecs(void)
+{
+  struct timespec ts;
+  clock_gettime(CLOCK_MONOTONIC, &ts);
+  return ((unsigned long long)ts.tv_sec * 1000000UL) + ts.tv_nsec/1000;
+}
+
+unsigned long long get_os_nsecs(void)
+{
+  struct timespec ts;
+
+  clock_gettime(CLOCK_MONOTONIC, &ts);
+  return ts.tv_sec * 1000000000UL + ts.tv_nsec;
+}
 */
 import "C"
 import (
 	"unsafe"
 
-	"github.com/flomesh-io/flb/pkg/bpf"
 	. "github.com/flomesh-io/flb/pkg/wq"
 )
 
@@ -41,7 +56,7 @@ func DpPolMod(w *PolDpWorkQ) int {
 		pa.ebs = C.uint(w.Ebs)
 		pa.tok_c = pa.cbs
 		pa.tok_e = pa.ebs
-		pa.lastc_uts = bpf.GetOsUSecs()
+		pa.lastc_uts = C.get_os_usecs()
 		pa.laste_uts = pa.toksc_pus
 		pa.drop_prio = FLB_PIPE_COL_YELLOW
